pyproject.toml
[tool.poetry]
name = "vote-gpt"
version = "0.1.0"
description = ""
authors = ["tallywiesenberg <t.j.wies@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.2"
google-api-python-client = "^2.130.0"
python-dotenv = "^1.0.1"
transformers = "^4.41.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
